issue-2725: mv acquire release logic from DR
