+++ cloud/blockstore/libs/storage/disk_registry/disk_registry_state.h
@@ -447,6 +447,7 @@ public:
 
     NProto::TError GetDiskInfo(const TDiskId& diskId, TDiskInfo& diskInfo) const;
     NProto::EDiskState GetDiskState(const TDiskId& diskId) const;
+    const TDiskState* GetDisk(const TDiskId& diskId) const;
     NProto::TError GetShadowDiskId(
         const TDiskId& sourceDiskId,
         const TCheckpointId& checkpointId,
